Add feature toggles for gram backend and cli interface
If I want to pull this in as a library to add my own backend,
I don't necessarily want to pull in the gram backend and the clap
interface to reduce the size of my dependencies.

By default, everything is enabled, nothing should change for the
"cargo run" use-case.

With this I can declare the dependency as
```
gqlite = { path = "../gqlite", default-features = false }
```