Reduce amount of pointer indirections
The PreparedStatement now consumes itself on calling run, i.e. it can only be run once.
This allows us to move the plan into the cursor and not have to keep a `Rc` copy of it.
We can now put the plan behind simple `Box`es, the `Rc<RefCell<_>>` is no longer needed.

Using associated types on Backend/Statement and a declared type parameter on the Cursor
allows us to remove some more trait object usages. It is really only the signature for
`Database::run` that had to suffer a bit, but it can be managed with the help if some
type aliases.