Add basic WHERE support for planner
I'm trying to pass the WhereAcceptance.feature, which is why this
includes several disparate things:

- Basic where clause parsing
- Parsing basic binary logic
- Moving the Predicate concepts over to Expr

The planner currently just tacks the predicates in WHERE after it's
solved all the expands, which will have catastrophic performance vs
early evaluation of predicates, but it's a good start. We'll want to
explore "hoisting" predicates up into the plan tree to prune stuff
as early as possible.. but that's likely best done guided by some
benchmarks.