Disconnect frontend from Val
Val is bad, because it mandates a heap layout for potentially huge
allocations (either by single values being large, like big strings,
or by lots and lots of values). We've made this mistake in the past;
it is much better if the backend owns how values are laid out in memory,
because that allows it to use the same layout as each backend chooses
to use on disk.

This has another nice side-effect in that it makes the frontend more
self-contained, allowing it to be reused for other fun purposes.