Weekend project: Clean up public rust API
I kind of don't like any of what is happening here, but it's a step
towards more understanding at least.

The root driver is that I want to pass the WHERE spec, which requires
being able to project whole nodes. Our "Val" type has a node, but it's
like.. it's the internal node inside the gram engine, just some ints.

This then leads down a rabbit hole of.. how do we expose the projected
row from a cypher query in an efficient way? Especially taking into
account that we presumably would like it to cross boundaries into other
languages (eg. mapping to python, javascript and so on).

Alongside that, you've got this desire that the backends are allowed to
own how values - at least the variably sized ones - are laid out. If
someone wants to hack up a backend that can deal with 10GiB string
properties without issues, or that has some really clever ways to
optimize lists of ints or whatnot.. basically you don't want the
backend to be forced to use the values the API exposes.

That last desire points to indirection or abstraction. Competing with
that you have a desire to limit allocation. You don't want an API that
forces you to allocate space for each string you return in the result
iteration hot loop.

Long story short: This ultimately is about getting rid of the weird
Rel and Node representations in the public API, and to clean it up
a bit in general. That comes at the cost of making the gram backend
messier..