Bulk hacking on WITH statement
Working towards getting more of the spec passing; my general thought
process here is that I can be as messy as I want in the gram backend,
rather messy in the cucumber tests, experimental in the rust API and
a bit tidier in the planner.

The planner is still very.. pragmatically written? But either way;
the point here is to get much of the spec green to get a vibe for
what needs to work and where there are dark secrets.

Once it's "correct" from the TCKs point of view, we can iterate on
architecture, performance and so on with much more confidence we're
not painting ourselves into a corner.